refactor(e2e): extract tokenization utilities
